fix(extension): bound every request to the PingCRM instance
3ea8e2e diagnosed the starvation correctly but fixed one of eleven call sites.
Chrome allows ~6 concurrent connections per host, so a hung background request
starves the PingCRM tab's own API calls. That is user-visible as a dashboard
rendering "0 contacts" on an account with thousands, because its requests never
leave the browser: nothing reaches Caddy or the backend, and the same page loads
fine in a profile where the extension is not running.

pairing.js had a timeout. These ten did not, all against ${apiUrl}:

  service-worker.js  suggestions, linkedin/push x2, extension/pair, regenerate
  sync.js            linkedin/push, extension/refresh
  sync-facebook.js   meta/push
  sync-instagram.js  meta/push
  twitter-sync.js    integrations/twitter/cookies

Rather than repeat AbortSignal.timeout() at each site and rely on the next
caller remembering, they now share apiFetch(), which applies a 30s timeout and
leaves an explicit caller signal alone. A test asserts no bare fetch() to
${apiUrl} remains, so a new call site cannot quietly reintroduce this.

service-worker.js was 663 lines and had to be split to be edited. Handlers moved
to handlers-{linkedin,suggestions,meta,account}.js, with the cache, badge,
throttled auto-syncs, and backfill in their own modules. What remains is the
dispatch table. The router now also catches a handler rejecting before it
responds, which previously left the sender waiting for a port that never
answered.

Behaviour is unchanged: the existing 27 tests still pass, including the router
and PROFILE_VISIT cases that exercise the extracted handlers end to end.
